backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
apscheduler
dateparser
python-dotenv
email-validator
